backend/requirements.txt
pyright
pydantic
fastapi[standard]
di[anyio]
pyjwt
passlib
